+++ RenderSystems/Metal/src/Vao/OgreMetalStagingBuffer.mm
@@ -373,7 +373,7 @@ namespace Ogre
         //Metal has alignment restrictions of 4 bytes for offset and size in copyFromBuffer
         size_t freeRegionOffset = getFreeDownloadRegion( alignToNextMultiple( srcLength, 4u ) );
 
-        if( freeRegionOffset == -1 )
+        if( freeRegionOffset == (size_t)(-1) )
         {
             OGRE_EXCEPT( Exception::ERR_INVALIDPARAMS,
                          "Cannot download the request amount of " +
@@ -446,7 +446,7 @@ namespace Ogre
         //Metal has alignment restrictions of 4 bytes for offset and size in copyFromBuffer
         size_t freeRegionOffset = getFreeDownloadRegion( alignToNextMultiple( srcLength, 4u ) );
 
-        if( freeRegionOffset == -1 )
+        if( freeRegionOffset == (size_t)(-1) )
         {
             OGRE_EXCEPT( Exception::ERR_INVALIDPARAMS,
                          "Cannot download the request amount of " +
